+++ examples/provider/provider.tf
@@ -1,3 +1,12 @@
-provider "scaffolding" {
-  # example configuration here
+terraform {
+  required_providers {
+    awscontrib = {
+      source = "hans-m-song/awscontrib"
+    }
+  }
+}
+
+provider "awscontrib" {
+  profile = "example"
+  region  = "ap-southeast-2"
 }